requirements.txt
Django~=5.1.4
scour~=0.38.2
pillow~=11.1.0